+++ Stardust.Data/Monitors/跟踪项.cs
@@ -107,13 +107,13 @@ namespace Stardust.Data.Monitors
         [BindColumn("Total", "总次数。累计埋点采样次数", "")]
         public Int64 Total { get => _Total; set { if (OnPropertyChanging("Total", value)) { _Total = value; OnPropertyChanged("Total"); } } }
 
-        private Int32 _Errors;
+        private Int64 _Errors;
         /// <summary>错误数</summary>
         [DisplayName("错误数")]
         [Description("错误数")]
         [DataObjectField(false, false, false, 0)]
         [BindColumn("Errors", "错误数", "")]
-        public Int32 Errors { get => _Errors; set { if (OnPropertyChanging("Errors", value)) { _Errors = value; OnPropertyChanged("Errors"); } } }
+        public Int64 Errors { get => _Errors; set { if (OnPropertyChanging("Errors", value)) { _Errors = value; OnPropertyChanged("Errors"); } } }
 
         private Int32 _Cost;
         /// <summary>平均耗时。总耗时除以总次数，单位毫秒</summary>
@@ -275,7 +275,7 @@ namespace Stardust.Data.Monitors
                     case "Timeout": _Timeout = value.ToInt(); break;
                     case "Days": _Days = value.ToInt(); break;
                     case "Total": _Total = value.ToLong(); break;
-                    case "Errors": _Errors = value.ToInt(); break;
+                    case "Errors": _Errors = value.ToLong(); break;
                     case "Cost": _Cost = value.ToInt(); break;
                     case "AlarmThreshold": _AlarmThreshold = value.ToInt(); break;
                     case "AlarmErrorRate": _AlarmErrorRate = value.ToDouble(); break;
